src/api/requirements.txt
uvicorn==0.17.5
python-dotenv==0.20.0
fastapi==0.75.1
pydantic==1.8.2
SQLAlchemy==1.4.28
python_dateutil==2.8.2